Bugfix ScroogeUser
