+++ src/ralph_scrooge/models/__init__.py
@@ -10,13 +10,13 @@ from ralph_scrooge.models.extra_cost import (
     ExtraCostType,
     SupportCost,
 )
-from ralph_scrooge.models.user import ScroogeUser
 from ralph_scrooge.models.owner import (
     OwnershipType,
     Owner,
     ServiceOwnership,
     TeamManager,
     UserProfile,
+    ScroogeUser
 )
 
 from ralph_scrooge.models.pricing_object import (
