+++ src/ralph_scrooge/tests/rest/test_costcard.py
@@ -20,7 +20,9 @@ from ralph_scrooge.tests.utils import factory
 
 class TestCardCost(TestCase):
     def setUp(self):
-        get_user_model().objects.create_superuser('test', 'test@test.test', 'test')
+        get_user_model().objects.create_superuser(
+            'test', 'test@test.test', 'test'
+        )
         self.client = APIClient()
         self.client.login(username='test', password='test')
